Add output row and column labels
